Tilemap Added
Used the TileMap and downloaded the sprite and made the map for our game. seocndly added a new layer Top Layer which contains all other objects and where player will interact. now on the second layer added the tilema collider which adds collider to each of the tile but to make sure that it adds the collider to whole object we added a composite coliider 2d which made sure that are coliider is added to whole one object rather than single itself. and upon adding the composite collider it also adds a rigid body so we need to make sure that we turn the rigidbody into static else our 2d scene will fall off.